+++ laim/laim.py
@@ -17,21 +17,20 @@ from .util import drop_privileges, unfold
 from .log import log, format_message_structure
 from ._version import __version__
 
-
-TaskArguments = namedtuple('TaskArguments', 'sender recipients data')
+TaskArguments = namedtuple("TaskArguments", "sender recipients data")
 
 
 class Laim:
 
     def __init__(
-            self,
-            port=25,
-            user='laim',
-            max_queue_size=50,
-            config_file='/etc/laim/config.yml',
-            smtp_kwargs=None,
+        self,
+        port=25,
+        user="laim",
+        max_queue_size=50,
+        config_file="/etc/laim/config.yml",
+        smtp_kwargs=None,
     ):
-        setproctitle.setproctitle('laim')
+        setproctitle.setproctitle("laim")
         self.queue = queue.Queue(max_queue_size)
         self.notifier = sdnotify.SystemdNotifier()
         handler = LaimHandler(self.queue)
@@ -47,16 +46,19 @@ class Laim:
         self.controller.start()
 
         # Let systemd we're done binding to the network socket
-        self.notifier.notify('READY=1')
-        log({
-            'action': 'started',
-            'port': port,
-            'max_queue_size': max_queue_size,
-            'config_file': config_file,
-            'py': platform.python_version(),
-        }, sender=self)
+        self.notifier.notify("READY=1")
+        log(
+            {
+                "action": "started",
+                "port": port,
+                "max_queue_size": max_queue_size,
+                "config_file": config_file,
+                "py": platform.python_version(),
+            },
+            sender=self,
+        )
 
-        with open(config_file, 'r') as config_fh:
+        with open(config_file, encoding="utf-8") as config_fh:
             drop_privileges(user)
             self.config = yaml.safe_load(config_fh)
 
@@ -67,15 +69,13 @@ class Laim:
         # Allow client sessions to be created
         privilege_event.set()
 
-
     def handle_message(self, sender, recipients, message):
-        raise NotImplementedError('The handler must implement handle_message')
-
+        raise NotImplementedError("The handler must implement handle_message")
 
     def run(self):
         worker_thread = threading.Thread(
             target=self._start_worker,
-            name='Laim worker',
+            name="Laim worker",
             daemon=False,
         )
         worker_thread.start()
@@ -83,15 +83,12 @@ class Laim:
         self.stop_event.wait()
         self.controller.stop()
 
-
     def stop(self):
         self.queue.put(None)
         self.stop_event.set()
 
-
-    def _signalhandler(self, signum, frame): # pylint: disable=unused-argument
+    def _signalhandler(self, signum, frame):  # pylint: disable=unused-argument
         self.stop()
-
 
     def _start_worker(self):
         while True:
@@ -100,35 +97,37 @@ class Laim:
                 break
 
             start_time = time.time()
-            message_string = task_args.data.decode('utf-8')
+            message_string = task_args.data.decode("utf-8")
             message = message_from_string(message_string)
-            raw_subject = message.get('subject')
+            raw_subject = message.get("subject")
 
             # Decode the subject to make it easier to consume for handlers
             decoded_subject = None
             if raw_subject:
                 decoded_subject = str(make_header(decode_header(raw_subject)))
-                message.replace_header('subject', decoded_subject)
+                message.replace_header("subject", decoded_subject)
 
             log_data = {
-                'action': 'handle-message',
-                'parse_time': '%.3fs' % (time.time() - start_time),
-                'sender': task_args.sender,
-                'recipients': ','.join(task_args.recipients),
-                'msg_structure': format_message_structure(message),
-                'raw_subject': unfold(raw_subject),
-                'decoded_subject': decoded_subject,
-                'msg_defects': ','.join(e.__class__.__name__ for e in message.defects),
+                "action": "handle-message",
+                "parse_time": f"{time.time() - start_time:.3f}s",
+                "sender": task_args.sender,
+                "recipients": ",".join(task_args.recipients),
+                "msg_structure": format_message_structure(message),
+                "raw_subject": unfold(raw_subject),
+                "decoded_subject": decoded_subject,
+                "msg_defects": ",".join(e.__class__.__name__ for e in message.defects),
             }
 
             try:
-                handler_data = self.handle_message(task_args.sender, task_args.recipients, message)
+                handler_data = self.handle_message(
+                    task_args.sender, task_args.recipients, message
+                )
                 if handler_data:
                     log_data.update(handler_data)
-            except Exception as ex: # pylint: disable=broad-except
-                log_data['action'] = 'handle-message-error'
-                log_data['error'] = ex.__class__.__name__
-                log_data['error_msg'] = str(ex)
+            except Exception as ex:  # pylint: disable=broad-except
+                log_data["action"] = "handle-message-error"
+                log_data["error"] = ex.__class__.__name__
+                log_data["error_msg"] = str(ex)
             finally:
                 log(log_data, start_time, sender=self)
 
@@ -137,41 +136,41 @@ class LaimHandler:
     def __init__(self, task_queue):
         self.task_queue = task_queue
 
-
-    async def handle_DATA(self, server, session, envelope): # pylint: disable=invalid-name,unused-argument
+    async def handle_DATA(
+        self, server, session, envelope
+    ):  # pylint: disable=invalid-name,unused-argument
         start_time = time.time()
         mail_from = envelope.mail_from
         data = envelope.content
         recipients = envelope.rcpt_tos
         log_data = {
-            'action': 'queued-message',
-            'mail_from': mail_from,
-            'recipients': ','.join(recipients),
-            'msg_size': len(data),
-            'client': session.host_name,
+            "action": "queued-message",
+            "mail_from": mail_from,
+            "recipients": ",".join(recipients),
+            "msg_size": len(data),
+            "client": session.host_name,
         }
         try:
             self.task_queue.put_nowait(TaskArguments(mail_from, recipients, data))
             log(log_data, start_time, sender=self)
         except queue.Full:
-            log_data['action'] = 'queue-full'
+            log_data["action"] = "queue-full"
             log(log_data, start_time, sender=self)
-            return '552 Exceeded storage allocation'
+            return "552 Exceeded storage allocation"
 
-        return '250 OK'
+        return "250 OK"
 
 
 class LaimController(Controller):
     def __init__(self, privilege_event, handler, port, smtp_kwargs=None):
-        super().__init__(handler, hostname='localhost', port=port)
+        super().__init__(handler, hostname="localhost", port=port)
         self.smtp_kwargs = smtp_kwargs
         self.privilege_event = privilege_event
 
-
     def factory(self):
         kwargs = {
-            'enable_SMTPUTF8': True,
-            'ident': 'laim %s' % __version__,
+            "enable_SMTPUTF8": True,
+            "ident": f"laim {__version__}",
         }
         if self.smtp_kwargs:
             kwargs.update(self.smtp_kwargs)
@@ -180,14 +179,14 @@ class LaimController(Controller):
 
 
 class LaimSMTP(SMTP):
-    '''
+    """
     Subclass to make sure no sessions are created before we've dropped
     privileges.
-    '''
+    """
+
     def __init__(self, privilege_event, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.privilege_event = privilege_event
-
 
     def _create_session(self):
         self.privilege_event.wait()