+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
+UNRELASED -
+------------------
+
+### Removed
+- Debian packages for bullseye.
+
+### Added
+- Debian packages for trixie.
+
+
 1.2.0 - 2025-03-18
 ------------------